# Optimize visual quality of converted ugoira