The current ffmpeg paras for converting have some issues:
E.g. apng is a lossless format only, but the current default setting added a denoiser that decreased visual quality and cost more time and increased file size.
E.g. the -r (Set frame rate, can apply to input or output) option is useless and incorrect. because the https://trac.ffmpeg.org/wiki/Concatenate#demuxer creates a variable frame rate input. so this option should never use in most time.

Test used ugoira: https://www.pixiv.net/en/artworks/46281014

Changes:
the .webp now uses lossy encoding with only the least quality loss. the quality is only behind the lossless one.
the .webm (uses vp9) now uses a better bitrate control method, than ffmpeg's default Average Bitrate with a very low 256k.

v20221029 settings:
.png
size: 22,006KB, cost time: 3s
.webm
size: 258KB, cost time: 9s
.webp
size: 1,820KB, cost time: 5s

New settings:
.png
size: 21,667KB, cost time: 2s
.webm
size: 910KB, cost time: 9s
.webp
size: 3,809KB, cost time: 7s

New settings but -lossless 1:
.png
size: 21,667KB, cost time: 2s
.webm
size: 7,045KB, cost time: 10s
.webp
size: 10,420KB, cost time: 10s

The source .zip size: 3,532KB
The original .ugoria size: 3,534KB