@Nandaka 2022-11-22
You still need to set the `-r 999`, else the max frame rate will be limited to 25fps.

I'm using https://www.pixiv.net/en/artworks/102233145 as the base and changed the delay manually and the new params failed to create the mp4 with variable frame rate.

modified [ugoira](https://www.mediafire.com/file/4j41xdoy2bza27e/102233145_ugoira1920x1080_-_Harucow%25F0%259F%2590%2584%25F0%259F%25A5%259B.ugoira/file)
result [old param](https://www.mediafire.com/file/o44g8bewvr049vm/old_params.zip/file)
result [new param](https://www.mediafire.com/file/o44g8bewvr049vm/old_params.zip/file)

```json
{"src":"https://i.pximg.net/img-zip-ugoira/img/2022/10/26/00/00/56/102233145_ugoira1920x1080.zip","originalSrc":"https://i.pximg.net/img-zip-ugoira/img/2022/10/26/00/00/56/102233145_ugoira1920x1080.zip","mime_type":"image/jpeg","frames":[{"file":"000000.jpg","delay":1},{"file":"000001.jpg","delay":1},{"file":"000002.jpg","delay":12},{"file":"000003.jpg","delay":12},{"file":"000004.jpg","delay":1},{"file":"000005.jpg","delay":1},{"file":"000006.jpg","delay":12},{"file":"000007.jpg","delay":12},{"file":"000008.jpg","delay":1},{"file":"000009.jpg","delay":1},{"file":"000010.jpg","delay":12},{"file":"000011.jpg","delay":12},{"file":"000012.jpg","delay":1},{"file":"000013.jpg","delay":1},{"file":"000014.jpg","delay":12},{"file":"000015.jpg","delay":12},{"file":"000016.jpg","delay":1},{"file":"000017.jpg","delay":1},{"file":"000018.jpg","delay":12},{"file":"000019.jpg","delay":12},{"file":"000020.jpg","delay":1},{"file":"000021.jpg","delay":1},{"file":"000022.jpg","delay":12},{"file":"000023.jpg","delay":12}],"zipSize":3008946}
```

```ini
[FFmpeg]
ffmpeg = .\ffmpeg.exe
ffmpegCodec = libvpx-vp9
ffmpegExt = mp4
ffmpegParam = -row-mt 1 -deadline good -crf 20 -vsync 2 -r 999 -pix_fmt yuv420p
webpCodec = libwebp
webpParam = -row-mt 1 -lossless 0 -q:v 90 -loop 0 -vsync 2 -r 999
gifParam = -filter_complex [0:v]split[a][b];[a]palettegen=stats_mode=diff[p];[b][p]paletteuse=dither=bayer:bayer_scale=5:diff_mode=rectangle
apngParam = -vsync 2 -r 999 -plays 0
verboseOutput = False
```