@Nandaka 2022-11-22
apng default still need to be changed, because looks like got issue with new ffmpeg