@NewUserHa 2022-11-22
yes, there's an issue when the duration between images is like 1 ms (which is a rare case though).

> the max frame rate will be limited to 25fps.

no. The output plays at the speed that the way you modified. since you can see it play very faster than the original.
The output webm is not vfr, which may be that webm doesn't support vfr probably. but if you change it to mp4 container, you can see the vfr flag in its metadata.

But the real issue I found is that without '-r' it will drop frames. More details can see:
https://superuser.com/questions/908295/ffmpeg-libx264-how-to-specify-a-variable-frame-rate-but-with-a-maximum

to verify use:
```py
with Image.open(r"") as im:
  print(im.n_frames)
```

The source file has total 24 frames. but in the outputs some frames are dropped:
without '-r', the .webm, .webp, .png each have 6 frames, while with'-r' they have each have correct 24+1 frames (+1 is a fix for ffmpeg quirk).
I found, however, the gif is only 10 frames already from old versions.

So I think it is better to completely solve that with a new method, which is '-vsync passthrough'. But there will be some friendly warning info from ffmpeg though.