@cglmrfreeman 2022-11-22
At some point we should look into implementing [gifski ](https://github.com/ImageOptim/gifski)as an option for ugoira conversion too.