@NewUserHa 2022-11-22
gif is a lossy format and an old (low compression)format.

Because the source file is a sequence of .jpgs, I may choose to use another way to view instead of using any converting.