@Nandaka 2022-11-23
should we change the `-vsync 0` to `-fps_mode passthrough`, then? Considering it is going to be deprecated in the future?

From: https://ffmpeg.org/ffmpeg.html
```
-vsync parameter (global)
-fps_mode[:stream_specifier] parameter (output,per-stream)

    Set video sync method / framerate mode. vsync is applied to all output video streams but can be overridden for a stream by setting fps_mode. vsync is deprecated and will be removed in the future.

    For compatibility reasons some of the values for vsync can be specified as numbers (shown in parentheses in the following table). 
```

as for gif encoding, I don't want to support it directly. You can run post-processing cmd from `config.ini` by setting `enablePostProcessing = True` and configure the cmd in `postProcessingCmd`