@NewUserHa 2022-11-23
> should we change the -vsync 0 to -fps_mode passthrough?

ffmpeg 5.0 has no `-fps_mode`, even it is a pretty new version and released in recent months.
So using `-vsync` should be better. I don't think ffmpeg will remove `-vsync` any soon.


> apng default still need to be changed, because looks like got issue with new ffmpeg

Does the apng default now have any issue?