@Nandaka 2022-11-23
looks ok with `-plays 0 -vsync 0`